Separate Game into its own module
