+++ src/engine/mod.rs
@@ -2,4 +2,5 @@ mod shell;
 mod tank;
 
 pub mod context;
+pub mod game;
 pub mod world;
